Added Fee Module

+++ src/main/java/com/collegeManagement/app/entity/StudentFeeEntity.java
@@ -46,7 +46,7 @@ public class StudentFeeEntity {
     @Getter
     @Setter
     @Column(name ="ReceiptFK")
-    private double receiptFK;
+    private long receiptFK;
 
     @Getter
     @Setter
